+++ deno.json
@@ -1,10 +1,10 @@
 {
   "tasks": {
     "dev": "deno run --watch --allow-net main.ts",
-    "node1": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net main.ts --port 5801 --init",
-    "node2": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net main.ts --port 5802 --join http://localhost:5801",
-    "node3": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net main.ts --port 5803 --join http://localhost:5802",
-    "wallet": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ main.ts --wallet"
+    "node1": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net src/main.ts --port 5801 --init",
+    "node2": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net src/main.ts --port 5802 --join http://localhost:5801",
+    "node3": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ --allow-net src/main.ts --port 5803 --join http://localhost:5802",
+    "wallet": "deno run --watch --allow-read=user-files/ --allow-write=user-files/ src/main.ts --wallet"
   },
   "imports": {
     "@oak/oak": "jsr:@oak/oak@^17.1.1",
